+++ Development/MyApp.swift
@@ -7,6 +7,13 @@ import Compatibility
 @available(iOS 15, macOS 12, tvOS 17, watchOS 8, *)
 @main
 struct MyApp: App {
+    init() {
+        Application.track() // ensures Application.main.isFirstRun and Application.main.versions variables are properly set.
+        if Application.main.isFirstRun {
+            debug("First Run!")
+        }
+        debug("All versions run: \(Application.main.versionsRun)")
+    }
     var body: some Scene {
         WindowGroup {
             CompatibilityDemoView()