+++ Sources/Core/Test.swift
@@ -52,10 +52,7 @@ public final class Test: ObservableObject {
     }
     public let title: String
     public var task: TestClosure
-    #if canImport(Combine)
-    @Published
-    #endif
-    public var progress: TestProgress = .notStarted
+    @Published public var progress: TestProgress = .notStarted
     
     public init(_ title: String, _ task: @escaping TestClosure ) {
         self.title = title